requirements.txt
psycopg2
groq
requests
fastapi
sqlalchemy
python-dotenv
pillow
pymupdf